임의문서 불러오기 , 검색 기능 및 무한 리사이클러뷰 구현

+++ app/src/main/java/com/supremehyo/awiki/model/WikiModelImple.kt
@@ -8,29 +8,65 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject
 import javax.inject.Singleton
 
+//여기 들어올때 local에 쓰일지 api 통신을 할지 flag 를 context랑 같이 넘겨서
+//거기에 맞게 분기해서 필요에 맞게 통신하면 될듯함.
+
 @Singleton // 싱글톤으로 어디서 여러번 호출해도 같은 객체로 인식하도록 집어넣어진다.
 class WikiModelImple @Inject constructor ( //Inject는 힐트가 생성자를 알아볼 수 있도록 명시
     @ApplicationContext private val context : Context) : WikiModel{ // context를 hilt 인자로 넘길때 ApplicationContext 가 필요
 
 
-    override suspend fun insertWiki(dto: WikiContract) {
-        val wikidao = WikiDatabase.getInstance(context)!!.contactsDao()
-      var aaaa =  wikidao.insertWiki(dto)
-        Log.v("wegwadb" , aaaa.toString())
+    override suspend fun insertWiki(dto: WikiContract , localOrApi : String) {
+        when(localOrApi){
+            "local" ->{
+                val wikidao = WikiDatabase.getInstance(context)!!.contactsDao()
+                var aaaa =  wikidao.insertWiki(dto)
+            }"api" ->{
+              val wikidao = WikiDatabase.getInstance(context)!!.contactsDao()
+                var aaaa =  wikidao.insertWiki(dto)
+            }
+        }
+    }
+    override suspend fun getOneWiki(title: String , context  : Context , localOrApi : String): WikiContract {
+        if(localOrApi == "local"){
+            val localWikidao = WikiDatabase.getInstance(context)!!.contactsDao()
+            return localWikidao.getOneWiki(title)
+        }else{
+            val apiWikidao = WikiDatabase.getInstance(context)!!.contactsDao()
+            return apiWikidao.getOneWiki(title)
+        }
+    }
+    override suspend fun editWiki(dto: WikiContract , localOrApi : String) {
+
+        when(localOrApi){
+                "local" ->{
+                    val wikidao = WikiDatabase.getInstance(context)!!.contactsDao()
+                    wikidao.editWiki(dto)
+                }"api" ->{
+                    val wikidao = WikiDatabase.getInstance(context)!!.contactsDao()
+                    wikidao.editWiki(dto)
+                }
+        }
+    }
+    override suspend fun deleteWiki(long: Long , localOrApi : String) {
+        when(localOrApi){
+            "local" ->{
+                val wikidao = WikiDatabase.getInstance(context)!!.contactsDao()
+                wikidao.deleteWiki(long)
+            }"api" ->{
+                val wikidao = WikiDatabase.getInstance(context)!!.contactsDao()
+                wikidao.deleteWiki(long)
+        }
+        }
     }
 
-    override suspend fun getOneWiki(title: String , context  : Context): WikiContract {
+    override suspend fun randomGetOneWiki(long: Long): WikiContract {
         val wikidao = WikiDatabase.getInstance(context)!!.contactsDao()
-       return wikidao.getOneWiki(title)
+        return wikidao.randomGetOneWiki(long)
     }
 
-    override suspend fun editWiki(dto: WikiContract) {
+    override suspend fun getWikiListBySearch(title: String): List<WikiContract> {
         val wikidao = WikiDatabase.getInstance(context)!!.contactsDao()
-        wikidao.editWiki(dto)
-    }
-
-    override suspend fun deleteWiki(long: Long) {
-        val wikidao = WikiDatabase.getInstance(context)!!.contactsDao()
-        wikidao.deleteWiki(long)
+        return wikidao.getListWikiBySearch(title)
     }
 }
